# [util] improve performance of type check