Small change to improve performance of check whether is something a list of int by a factor of roughly 20:
Old implementation on list of `1000000` random ints:
```
861 ms ± 9.3 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
New implementation on list of `1000000` random ints:
```
48.3 ms ± 932 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
Fixes #1439