@marscher 2019-11-20
That's brilliant, did you come up with the groupby type idea? :+1: