@marscher 2019-11-20
the failed MSM tests are due to a change of version in openblas, I have reported it on Conda-forge (since it is actually on an increase in build number, not source code version) https://github.com/conda-forge/blas-feedstock/issues/40