@marscher 2019-11-20
The speed penalty of the old impl has to arise from the boolean and operations per element, not the type check. Since groupby reduces the boolean operations to the amount of types in the list, it is blazingly fast :zap: